lib/Compiler: Switch to standard JSON input Solidity API.
The change to standard JSON input is needed to align Solidity
compiler options for Embark and Embark-solc plugin.

The background of the change is to fix bug:

https://github.com/embark-framework/embark/issues/942

where the contract profiler was working for solcjs but not working with Embark-solc plugin
using binary solc release. The root cause was missing gasEstimate data in the compiled
contract JSON output data. The gasEstimate data is not available due to different compiler options
available for command line --combined-json and standard JSON input.

Gas estimation is command line solc compiler can be triggered by invoking command
line --gas option, the gas estimate is then printed out on the standard output but does not
end up in the Solidity compiler JSON output.

The solution to fix the problem is to use standard JSON input both for solcjs and solc compilers.
Additionally this change  aligns Embark and Embark-solc to use the same compiler options.

References: https://solidity.readthedocs.io/en/v0.4.25/using-the-compiler.html

Fixes: https://github.com/embark-framework/embark/issues/942